% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_icon.R
\name{open_icon}
\alias{open_icon}
\title{open_icon}
\usage{
open_icon(species)
}
\arguments{
\item{species}{name of species which corresponds with icon name}
}
\value{
returns an icon
}
\description{
This function opens a stored icon of an animal
}
