% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_species_info.R
\name{prepare_species_info}
\alias{prepare_species_info}
\title{prepare_species_info}
\usage{
prepare_species_info()
}
\value{
returns a list with character string variables containing descriptions for each species.
}
\description{
This function simply prepares information for the each species to be displayed in the shiny application. This function has no inputs.
}
