% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R, R/server.R
\name{server}
\alias{server}
\title{server}
\usage{
server(input, output)

server(input, output)
}
\value{
returns a server

returns a server
}
\description{
This function creates a server

This function creates a server
}
