% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemporalVisualisation-package.R
\docType{package}
\name{TemporalVisualisation-package}
\alias{TemporalVisualisation-package}
\title{TemporalVisualisation-package}
\description{
This package includes a subset of the Barro Colorado Forest Census Plot Data data.
Using this dataset, we will measure different forms of alpha, beta and gamma diversity for different values of the parameter q which gives weight to species richness compared to abundance (lower values of q increase the weight of species richness) in the calculation of species diversity.
}
\details{
\code{TemporalVisualisation} This package allows the user to model the stochastic nature of population growth and disease transmission.
Stochastic plots are compared with deterministic plots from previous practicals.
}
\author{
Dylan Carbone <Dylancarbone25@gmail.com> (maintainer) \cr
}
