% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemporalVisualisation-package.R
\docType{package}
\name{TemporalVisualisation-package}
\alias{TemporalVisualisation-package}
\title{TemporalVisualisation-package}
\description{
The following package allows the user to visualise changes in the distributions of urban wildife with time of day and day of survey. This package uses data collected during cameratrap surveys conducted by London HogWatch at ZSL.
}
\author{
Dylan Carbone <Dylancarbone25@gmail.com> (maintainer) \cr
}
